+++ src/mesh_utils/global_ids_updater_inline_impl.hh
@@ -43,8 +43,8 @@
 namespace akantu {
 
 /* -------------------------------------------------------------------------- */
-inline Int GlobalIdsUpdater::getNbData(const Array<Element> & elements,
-                                       const SynchronizationTag & tag) const {
+inline Int GlobalIdsUpdater::getNbData(const Array<Element> &elements,
+                                       const SynchronizationTag &tag) const {
   UInt size = 0;
   if (tag == SynchronizationTag::_giu_global_conn) {
     size += Mesh::getNbNodesPerElementList(elements) *
@@ -58,21 +58,21 @@ inline Int GlobalIdsUpdater::getNbData(const Array<Element> & elements,
 }
 
 /* -------------------------------------------------------------------------- */
-inline void GlobalIdsUpdater::packData(CommunicationBuffer & buffer,
-                                       const Array<Element> & elements,
-                                       const SynchronizationTag & tag) const {
+inline void GlobalIdsUpdater::packData(CommunicationBuffer &buffer,
+                                       const Array<Element> &elements,
+                                       const SynchronizationTag &tag) const {
   if (tag != SynchronizationTag::_giu_global_conn) {
     return;
   }
 
   int prank = mesh.getCommunicator().whoAmI();
 
-  const auto & global_nodes_ids = mesh.getGlobalNodesIds();
+  const auto &global_nodes_ids = mesh.getGlobalNodesIds();
   buffer << prank;
 
-  for (const auto & element : elements) {
+  for (const auto &element : elements) {
     /// get element connectivity
-    auto && current_conn =
+    auto &&current_conn =
         const_cast<const Mesh &>(mesh).getConnectivity(element);
 
     /// loop on all connectivity nodes
@@ -94,22 +94,22 @@ inline void GlobalIdsUpdater::packData(CommunicationBuffer & buffer,
 }
 
 /* -------------------------------------------------------------------------- */
-inline void GlobalIdsUpdater::unpackData(CommunicationBuffer & buffer,
-                                         const Array<Element> & elements,
-                                         const SynchronizationTag & tag) {
+inline void GlobalIdsUpdater::unpackData(CommunicationBuffer &buffer,
+                                         const Array<Element> &elements,
+                                         const SynchronizationTag &tag) {
   if (tag != SynchronizationTag::_giu_global_conn) {
     return;
   }
 
   MeshAccessor mesh_accessor(mesh);
-  auto & global_nodes_ids = mesh_accessor.getNodesGlobalIds();
+  auto &global_nodes_ids = mesh_accessor.getNodesGlobalIds();
 
   int proc;
   buffer >> proc;
 
-  for (const auto & element : elements) {
+  for (const auto &element : elements) {
     /// get element connectivity
-    auto && current_conn =
+    auto &&current_conn =
         const_cast<const Mesh &>(mesh).getConnectivity(element);
 
     /// loop on all connectivity nodes
@@ -131,8 +131,8 @@ inline void GlobalIdsUpdater::unpackData(CommunicationBuffer & buffer,
       }
 
       if (mesh.isSlaveNode(node)) {
-        auto & gid = global_nodes_ids(node);
-        AKANTU_DEBUG_ASSERT(gid == UInt(-1) or gid == index,
+        auto &gid = global_nodes_ids(node);
+        AKANTU_DEBUG_ASSERT(gid == -1 or gid == index,
                             "The node already has a global id, from proc "
                                 << proc << ", different from the one received "
                                 << gid << " " << index);
