+++ src/model/common/integration_scheme/integration_scheme.hh
@@ -58,7 +58,7 @@ public:
     _acceleration = 2,
   };
 
-  IntegrationScheme(DOFManager & dof_manager, const ID & dof_id, UInt order);
+  IntegrationScheme(DOFManager &dof_manager, const ID &dof_id, UInt order);
   ~IntegrationScheme() override = default;
 
   /* ------------------------------------------------------------------------ */
@@ -69,10 +69,10 @@ public:
   virtual void predictor(Real delta_t) = 0;
 
   /// generic interface of a corrector
-  virtual void corrector(const SolutionType & type, Real delta_t) = 0;
+  virtual void corrector(const SolutionType &type, Real delta_t) = 0;
 
   /// assemble the jacobian matrix
-  virtual void assembleJacobian(const SolutionType & type, Real delta_t) = 0;
+  virtual void assembleJacobian(const SolutionType &type, Real delta_t) = 0;
 
   /// assemble the residual
   virtual void assembleResidual(bool is_lumped) = 0;
@@ -98,16 +98,16 @@ public:
   /* ------------------------------------------------------------------------ */
 protected:
   /// The underlying DOFManager
-  DOFManager & dof_manager;
+  DOFManager &dof_manager;
 
   /// The id of the dof treated by this integration scheme.
   ID dof_id;
 
   /// The order of the integrator
-  UInt order;
+  Int order;
 
   /// last release of M matrix
-  UInt m_release{UInt(-1)};
+  Int m_release{-1};
 
   /// stores the values at begining of solve
   std::vector<std::unique_ptr<Array<Real>>> u_store;
@@ -116,8 +116,8 @@ protected:
 /* -------------------------------------------------------------------------- */
 // std::ostream & operator<<(std::ostream & stream,
 //                           const IntegrationScheme::SolutionType & type);
-std::istream & operator>>(std::istream & stream,
-                          IntegrationScheme::SolutionType & type);
+std::istream &operator>>(std::istream &stream,
+                         IntegrationScheme::SolutionType &type);
 /* -------------------------------------------------------------------------- */
 
 } // namespace akantu
